+++ src/main/java/thecodex6824/thaumcraftfix/mixin/vanilla/ModelBipedMixin.java
@@ -18,7 +18,7 @@
  *  along with Thaumcraft Fix.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package thecodex6824.thaumcraftfix.core.mixin.vanilla;
+package thecodex6824.thaumcraftfix.mixin.vanilla;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
