+++ src/main/resources/mixin/vanilla.json
@@ -1,5 +1,5 @@
 {
-  "package": "thecodex6824.thaumcraftfix.core.mixin.vanilla",
+  "package": "thecodex6824.thaumcraftfix.mixin.vanilla",
   "refmap": "${refmap}",
   "minVersion": "${mixinbooterminversion}",
   "compatibilityLevel": "JAVA_8",
